Renamed DynamicProgramming/utils.py to DynamicProgramming/core.py
